agregados test unitarios

+++ back/dto/dto_operation.go
@@ -2,7 +2,7 @@ package dto
 
 type OperationDto struct {
 	Id      int
-	Number1 float32 `json:"number1"`
-	Number2 float32 `json:"number2"`
-	Result  float32 `json:"result"`
+	Number1 float64 `json:"number1"`
+	Number2 float64 `json:"number2"`
+	Result  float64 `json:"result"`
 }
