+++ back/model/operation.go
@@ -2,9 +2,9 @@ package model
 
 type Operation struct {
 	Id      int
-	Number1 float32
-	Number2 float32
-	Result  float32
+	Number1 float64
+	Number2 float64
+	Result  float64
 }
 
 type Operations []Operation
