+++ back/service/repositories/client.go
@@ -6,5 +6,5 @@ import (
 
 type OperationClient interface {
 	GetOperationById(id int) (model.Operation, error)
-	InsertOperation(operation model.Operation) (float32, error)
+	InsertOperation(operation model.Operation) (float64, error)
 }
